# backend/static/swagger/swagger.json
{
	"swagger": "2.0",
	"info": {
		"title": "Invoicey backend API",
		"description": "Api that connects the inner workings of Invoicey application",
		"version": "1.0.0"
	},
	"tags": [
		{
			"name": "Users",
			"description": "The user application endpoints"
		},
		{
			"name": "Clients",
			"description": "The clients application endpoints"
		},
		{
			"name": "Invoices",
			"description": "The invoices application endpoints"
		},
		{
			"name": "Payments",
			"description": "Contains Payments "
		}
	],
	"basePath": "/",
	"schemes": [
		"http",
		"https"
	],
	"security": [
		{
			"bearerAuth": []
		}
	],
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"paths": {
		"/": {
			"get": {
				"summary": "Returns a csrf_token for all post requests",
				"produces":[
					"application/json"
				],
				"security": [
				],
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"required": true,
						"type": "string",
						"description": "Required header to validate",
						"properties": {
							"Authorization": "Bearer {Token}"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
							}
						}
					}
				}
			}
		},
		"/api/authenticate/": {
			"post": {			
				"summary": "Accepts User JSON Auth Data",
				"produces": [
					"application/json"
				],
				"consumes": [
					"application/json"
				],
				"tags": [
					"Users"
				],
				"security": [
				],
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"required": true,
						"type": "object",
						"properties": {
							"email": {
								"type": "string",
								"example": "test@email.com"
							},
							"password": {
								"type": "string",
								"example": "password"
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"description": "If posted data is oks",
									"items": {
										"data": {
											"type": "boolean",
											"example": true
										},
										"refresh_token": {
											"type": "string",
											"example": "jwt"
										},
										"authenticated": {
											"type": "boolean",
											"example": false
										},
										"valid": {
											"type": "boolean",
											"example": true
										},
										"level": {
											"type": "string",
											"example": "NOTSET"
										},
										"is_activated": {
											"type": "boolean",
											"example": false
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"description": "Not Authorzed Request",
									"items": {
									    "authenticated":{
									    	"type": "boolean",
									    	"example": false
										},
									    "data": {
									    	"type": "boolean",
									    	"example": false
									    },
										"level": {
											"type": "string",
											"example": "warning"
										},
										"message": {
											"type": "string",
											"example": "Invalid credentials"
										},
										"refresh_token": {
											"type": "string"
										},
										"valid": {
											"type": "boolean",
											"example": false
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/overview-data/": {
			"get": {
				"summary": "Gets the dashboard data",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"required": true,
						"type": "string",
						"description": "Required header to validate"
					},
					{
						"name": "Activated",
						"in": "header",
						"required":false,
						"type":"string"
					}
				],
				"responses": {
					"200": {
						"content": [
							"application/json"
						]
					}
				}
			}
		},
		"/api/register-user/": {
			"post": {
				"summary": "Creating User Application",
				"produces": [
					"application/json"	
				],
				"consumes": [
					"application/json"	
				],
				"tags": [
					"Users"
				],
				"security": [
				],
				"parameters":[
					{
						"name": "json",
						"in": "body",
						"type": "object",
						"required": true,
						"description": "register-user data",
						"properties": {
							"email": {
								"type": "string",
								"example": "example@email.com"
							},
							"name": {
								"type": "string",
								"example": "John. Doe"
							},
							"password": {
								"type": "string",
								"example": "password"
							},
							"busi_nm": {
								"type": "string",
								"example": "Business Name"
							}
						}
					}
				],
				"responses": {
					"200":{
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"items": {
										"type": "object",
										"properties": {
											"created": {
												"type":"boolean"
											},
											"message": {
												"type":"string"
											},
											"status": {
												"type":"string"
											}
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/all-client-data/": {
			"get": {
				"summary": "Returns all-clients data",
				"produces": [
					"application/json"
				],
				"tags": [
					"Clients"	
				],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"required": true
					},
					{
						"name": "Authorization",
						"in": "header",
						"type": "string",
						"required": true,
						"properties": {
							"Authorization": "Bearer {Token}"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "object",
												"properties": {
													"id": {
														"type": "string",
														"example": "EP{id}"
													},
													"name": {
														"type": "string",
														"example": "Test User Name"
													},
													"email": {
														"type": "string",
														"example": "email@email.com"
													},
													"phone": {
														"type": "string",
														"example": "9000000000",
														"length": 10
													},
													"birth_date": {
														"type": "string",
														"example": "dd/mm/yyyy"
													},
													"gender": {
														"type": "string",
														"example": "M",
														"length": 1
													}
												}
											}
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/invoices-data/": {
			"get": {
				"description": "Gets Invoices Data",
				"tags": [
					"Invoices"
				],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"required": true
					},
					{
						"name": "Authorization",
						"in": "header",
						"type": "string",
						"required": true,
						"properties": {
							"Authorization": "Bearer {Token}"
						}
					}
				],
				"responses": {
					"200": {
						"description": "A Successful Response",
						"content": [
							"application/json"
						]
					}
				}
			}
		},
		"/api/users/password-reset/": {
			"post": {
				"description": "Activate the token passed int",
				"tags": [
					"Users"
				],
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"required": true,
						"type": "object",
						"properties": {
							"email": {
								"type": "string",
								"example": "example@email.com"
							}
						}
					}
				],
				"security": [
				]
			}
		},
		"/api/users/activate/{token}/": {
			"post": {
				"description": "Activate the token passed int",
				"tags": [
					"Users"
				],
				"parameters": [
					{
						"name": "token",
						"in": "path",
						"required": true
					}
				],
				"responses":{
					"200":{
						"content":{
							"application/json":{
								
							}
						}
					}
				},
				"security": [
				]
			}
		},
		"/api/clients/": {
			"parameters": [
				{
					"name": "Authorization",
					"in": "header",
					"required": true,
					"type": "string",
					"properties": {
						"Authorization": "Bearer {Token}"
					}
				}
			],
			"post": {
				"description": "Creates the clients",
				"tags": [
					"Clients"
				],
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"required": true,
						"type": "object",
						"properties": {
							"name": {
								"type": "string",
								"example": "John. Client"
							},
							"email": {
								"type": "string",
								"example": "client@invoicey.com"
							},
							"birth_date": {
								"type": "string",
								"example": "12/11/2022"
							},
							"phone": {
								"type": "string",
								"example": "+234790989087"
							},
							"gender": {
								"type": "string",
								"example": "F/M"
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": [
							"application/json"
						]
					}
				}
			},
			"get": {
				"description": "Get the a single client with the given id",
				"tags": [
					"Clients"
				],
				"parameters": [
					{
						"name": "_id",
						"in": "query",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"content": [
							"application/json"
						]
					}
				}

			},
			"delete": {
				"description": "Deletes the client Data",
				"tags": [
					"Clients"
				],
				"parameters":[
					{
						"name":"_id",
						"in":"query",
						"required": true,
						"type": "integer"
					}
				],
				"responses":{
					"content":[
						"application/json"
					]
				}
			}
		},
		"/api/bank/": {
			"parameters":[
				{
					"name": "Authorization",
					"in": "header",
					"required":true,
					"type":"string"
				}
			],
			"get": {
				"description": "Get the user or Merchant Bank Accounts",
				"tags": [
					"Payments"
				],
				"responses":{
					"200": {
						"description": "OK",
						"content": {
							"application/json":{
								"schema": {
									"type": {
										"items": {
											"type": "object",
											"properties":{
												"acct_num":{

												},
												"bank_name": {

												},
												"bank_code": {

												},
												"acct_name": {

												},
												"first_name": {

												},
												"last_name": {

												}
											}
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"description": "Saves the user bank account details",
				"tags": [
					"Payments"
				],
				"parameters": [
					{
						"name": "body",
						"required": true,
						"in": "body",
						"type": "object",
						"properties": {
							"acct_num": {
								"type":"string"
							},
							"bank_name": {
								"type": "string",
								"example": "Bank Name"
							},
							"acct_name": {
								"type":"string",
								"example": "Account Name"
							}, 
							"bank_code": {
								"type": "string",
								"example": "099"
							}, 
							"first_name": {
								"type": "string",
								"example": "John"
							},
							"last_name": {
								"type": "string",
								"example": "Doe"
							},
							"other": {
								"type": "string",
								"example": "Arthur"
							}
						}
					}
				],
				"responses": {
					"content": [
						"application/json"
					]
				}
			},
			"delete": {
				"description": "Deletes the user bank account",
				"tags": [
					"Payments"
				],
				"responses": {
					"content": {
						"application/json": {

						}
					}
				}
			}
		},
		"/api/activate_required/": {
			"description": "Activates user based on authorization token",
			"post": {
				"tags": [
					"Users"
				],
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"required": true
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json":{
								"schema": {
									"type": "object",
									"items": {
										"message": {
											"type": "string",
											"example": "User Account Activated"
										},
										"status": {
											"type": "string",
											"example": "success"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/users/resend-activation-link/":{
			"post":{
				"description": "Resends activation link",
				"tags":[
					"Users"
				],
				"security":[
					
				],
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"required": true,
						"type": "object",
						"properties":{
							"email": {
								"type":"string",
								"example":"example@email.com"
							}
						}
					}
				],
				"responses": {
					"200": {
						"content":{
							"application/json":{

							}
						}
					},
					"401": {
						"content": {
							"application/json": {

							}
						}
					}
				}
			}
		},
		"/api/invoice/":{
			"get": {
				"tags": [
					"Invoices"
				],
				"description": "Gets a single Invoice passing in as a query param inv_id",
				"parameters": [
					{
						"name":"inv_id",
						"in": "query",
						"required":true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"Invoices"
				],
				"description": "Accepts a json body in the request",
				"parameters":[
					{
						"name": "body",
						"in":"body",
						"required": true,
						"properties":{
							"product_name": {
								"type": "string",
								"example": "Product Name"
							},
							"description": {
								"type": "string",
								"example": "Product Description"
							},
							"client_name": {
								"type": "string",
								"example": "Client Full Name"
							},
							"amount": {
								"type": "double",
								"example": 2000.00
							},
							"has_paid": {
								"type":  "boolean",
								"example": false
							},
							"py_type": {
								"type": "string",
								"example": "Payment type"
							},
							"due_date": {
								"type": "string",
								"example": "dd/mm/yyyy"
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {}
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Invoices"
				],
				"parameters":[
					{
						"name": "body",
						"in": "body",
						"required": true,
						"type": "object",
						"properties": {
							"inv_id": {
								"type": "string",
								"example": "invoice id"
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {}
								}
							}
						}
					}
				}
			},
			"parameters": [
				{
					"name":"Authorization",
					"in":"header",
					"required":true
				}
			]
		},
		"/api/multi/invoice/": {
			"get":{
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"required": true,
						"type":"integer"
					},
					{
						"name":"Authorization",
						"in":"header",
						"required":true,
						"type": "string"
					}
				],
				"description": "Get all clients based on current authenticated user",
				"tags": [
					"Invoices"
				],
				"responses":{
					"200":{
						"description":"OK"
					},
					"401": {
						"description":"Unauthorized"
					}
				}
			}
		},
		"/api/get-user-data/": {
			"get": {
				"description": "Gets the authenticated user data",
				"tags":[
					"Users"
				],
				"parameters":[
					{
						"name": "Authorization",
						"in": "header",
						"required":true,
						"type":"string"
					}
				],
				"responses":{
					"200":{
						"description": "OK",
						"summary": "returns the current user data",
						"content":{
							"application/json":{

							}
						}

					},
					"401":{

					}
				}
			}
		},
		"/api/update-transaction-status/": {
			"get": {
				"description": "Gets Transaction status",
				"parameters": [
					{ 
						"name": "Authorization",
						"in": "header",
						"required": true
					},
					{
						"name": "trsc_ref",
						"in": "query",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {

							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								
							}
						}
					}
				}
			}
		},
		"/update-transaction-status/": {
			"get": {
				"description": "Gets the transaction status",
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"required": true
					},
					{
						"name": "trsc_id",
						"in": "query",
						"required": true
					}
				],
				"tags": [
					"Payments"
				],
				"responses": {
					"200": {	
						"description": "OK",
						"content":{
							"application/json":{

							}
						}
					},
					"401": {

					}
				}
			}
		},
		"/send-trsc-link/": {
			"post": {
				"description": "Resends another trsc link",
				"tags":[
					"Payments"
				],
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"required": true
					},
					{
						"name": "body",
						"in": "body",
						"required": true,
						"type": "object",
						"properties": {
							"trsc_ref": {
								"type":"string",
								"example":"some long token" 
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json":{

							}
						}
					},
					"401": {

					}
				}
			}
		}
	}
}